Added transistion support
You can now have transistion effects when revealing/hiding items by
adding data attributes to the item telling the controller what classes
to add to the item for the transistion to happen. Got the information
about how to do this from the great articles and repos linked below so all
credit to the authors for their great work.

Sebastian De Deyne
https://sebastiandedeyne.com/javascript-framework-diet/enter-leave-transitions/

Mike McCall article on dev.to
https://dev.to/mmccall10/tailwind-enter-leave-transition-effects-with-stimulus-js-5hl7

The el-transistion package by Mike McCall
https://github.com/mmccall10/el-transition

And a pull request from tomclose on the el-transistion repo to fix an
issue with original classes being overwritten
https://github.com/mmccall10/el-transition/pull/4